samples/uds_new/client/requirements.txt
can-isotp==2.0.3
udsoncan==1.21.2